0.428  Feature: Per-class draw limits. Defaults are: owner unlimited, op unlimited, regular 4096 (16x16x16), guest 512 (8*8*8)  Feature: Per-class antigrief detection. Defaults are: owner unmetered, op umetered, regular 45/6 (7.5 blocks/second), guest 35/5 (7 blocks/second)  Feature: Memory usage of draw commands has been reduced by 6 bytes per block (37% improvement).  Fix: Fixed a bug with /undo that caused more than one command to be undone at once.  Fix: Fixed maps generated with "Realistic terrain" (MapGenerator) having an empty layer at the bottom of the map.  Change: Default classes have been tweaked:          - regular can now use draw commands          - op can now lock/unlock  Change: Improved arrangement of Permission list items.  Change: Removed obsolete ConfigKey variables AntigriefBlockCount and AntigriefBlockInterval

+++ fCraft/fCraft/World/BlockUpdate.cs
@@ -5,14 +5,14 @@ using System;
 namespace fCraft {
     public struct BlockUpdate {
         public Player origin;
-        public int x, y, h;
+        public short x, y, h;
         public byte type; 
 
         public BlockUpdate( Player _origin, int _x, int _y, int _h, byte _type ) {
             origin = _origin;
-            x = _x;
-            y = _y;
-            h = _h;
+            x = (short)_x;
+            y = (short)_y;
+            h = (short)_h;
             type = _type;
         }
     }
